Move `Competence` and `BlockedStep` to `compound` module
These types follow directly from the introduction of CompoundStep.
Moving them decouples code that depends on them from the `arraystep` module.
